backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
python-dotenv
passlib[bcrypt]==1.7.4
requests
PyJWT
psycopg2-binary
PyPDF2
python-docx
python-multipart
bcrypt==3.2.2